feature: added upvotes to create and search routes
