+++ src/handlers/card_handler.rs
@@ -1,24 +1,12 @@
 use actix_web::{web, HttpResponse};
-use qdrant_client::qdrant::point_id::PointIdOptions;
-use qdrant_client::qdrant::value::Kind;
-use qdrant_client::qdrant::with_payload_selector::SelectorOptions;
-use qdrant_client::qdrant::{PayloadIncludeSelector, PointStruct};
-use qdrant_client::{prelude::*, qdrant::WithPayloadSelector};
+use qdrant_client::qdrant::PointStruct;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 
-use crate::{errors::DefaultError, operators::card_operator::create_openai_embedding};
+use crate::operators::card_operator::{get_qdrant_connection, search_card_query};
+use crate::operators::card_operator::create_openai_embedding;
 
 use super::auth_handler::LoggedUser;
-
-pub async fn get_qdrant_connection() -> Result<QdrantClient, DefaultError> {
-    let qdrant_url = std::env::var("QDRANT_URL").expect("QDRANT_URL must be set");
-    QdrantClient::new(Some(QdrantClientConfig::from_url(qdrant_url.as_str())))
-        .await
-        .map_err(|_err| DefaultError {
-            message: "Failed to connect to Qdrant",
-        })
-}
 
 #[derive(Serialize, Deserialize)]
 pub struct CreateCardData {
@@ -30,7 +18,7 @@ pub struct CreateCardData {
 
 pub async fn create_card(
     card: web::Json<CreateCardData>,
-    user: LoggedUser,
+    user: Option<LoggedUser>,
 ) -> Result<HttpResponse, actix_web::Error> {
     let embedding_vector = create_openai_embedding(&card.content).await?;
 
@@ -38,15 +26,17 @@ pub async fn create_card(
         .await
         .map_err(|err| actix_web::error::ErrorBadRequest(err.message))?;
 
-    // add some evidence card in CSV form to the vector db (Create route) (Route takes the card [side, topic, author, date, content], creates an embedding for only content, puts content's vector into the db]
+    let id_str = user.map(|user| user.id.to_string()).unwrap_or("".to_string());
 
-    let payload: Payload = json!(
+    let payload: qdrant_client::prelude::Payload = json!(
         {
             "content": card.content.clone(),
             "topic": card.topic.clone(),
             "side": card.side.clone(),
-            "user_id": user.id.to_string(),
+            "user_id": id_str,
             "link": card.link,
+            "upvotes": 0,
+            "downvotes": 0,
             "created_at": chrono::Utc::now().to_rfc3339(),
         }
     )
@@ -67,88 +57,13 @@ pub struct SearchCardData {
     content: String,
 }
 
-#[derive(Serialize, Deserialize)]
-pub struct CardDTO {
-    id: String,
-    content: String,
-    side: String,
-    topic: String,
-    score: f32,
-    link: Option<String>,
-}
-
 pub async fn search_card(
     data: web::Json<SearchCardData>,
 ) -> Result<HttpResponse, actix_web::Error> {
     let embedding_vector = create_openai_embedding(&data.content).await?;
 
-    let qdrant = get_qdrant_connection()
-        .await
-        .map_err(|err| actix_web::error::ErrorBadRequest(err.message))?;
+    let cards = search_card_query(embedding_vector).await?;
 
-    let data = qdrant
-        .search_points(&SearchPoints {
-            collection_name: "debate_cards".to_string(),
-            vector: embedding_vector,
-            filter: None,
-            limit: 10,
-            with_payload: Some(WithPayloadSelector {
-                selector_options: Some(SelectorOptions::Include(PayloadIncludeSelector {
-                    fields: vec![
-                        "content".to_string(),
-                        "side".to_string(),
-                        "topic".to_string(),
-                        "user_id".to_string(),
-                        "link".to_string(),
-                    ],
-                })),
-            }),
-            with_vectors: None,
-            params: None,
-            score_threshold: None,
-            offset: None,
-            ..Default::default()
-        })
-        .await
-        .map_err(actix_web::error::ErrorBadRequest)?;
-
-    let cards: Vec<CardDTO> = data
-        .result
-        .iter()
-        .filter_map(|point| {
-            let id = match point.id.clone()?.point_id_options? {
-                PointIdOptions::Num(n) => n.to_string(),
-                PointIdOptions::Uuid(s) => s,
-            };
-            let content = point.payload.get("content")?;
-            let side = point.payload.get("side")?;
-            let topic = point.payload.get("topic")?;
-            let score = point.score;
-            let link = point.payload.get("link").and_then(|link| {
-                if let Some(Kind::StringValue(s)) = &link.kind {
-                    Some(s.clone())
-                } else {
-                    None
-                }
-            });
-
-            match (&content.kind, &side.kind, &topic.kind) {
-                (
-                    Some(Kind::StringValue(content)),
-                    Some(Kind::StringValue(side)),
-                    Some(Kind::StringValue(topic)),
-                ) => Some(CardDTO {
-                    id,
-                    content: content.clone(),
-                    side: side.clone(),
-                    topic: topic.clone(),
-                    score,
-                    link,
-                }),
-                (_, _, _) => None,
-            }
-        })
-        .collect();
 
     Ok(HttpResponse::Ok().json(cards))
 }
